Drop the shell_out logging down to trace level
This has to have been an oversight when thom created the trace
level, since live stdout logging produces a lot of output.

We also need this for adding live stdout logging for ohai which
produces a lot of output, where trace is certainly more
appropriate and makes the debug level difficult to read.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>