Stop doing weird things with STDOUT
This means we get output when going to logfiles, while daemonized or
when sending to pipes.  Piping trace output to grep/less is why I
consider the current behavior broken and this to be a bugfix.

Since this is all debug/trace logging tweaking this should not
be impacting anyone's running systems (https://xkcd.com/1172/).

Under ideal conditions we should probably figure out some way to send
output through the logger when trace is enabled and there's no STDOUT,
but this is probably a VERY unlikely use case and is just not super
easy.  And daemonized trace logging of live stdout to a logfile will
actually work correctly if someone is using something like systemd and
just using STDOUT logging like it is the 21st century.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>